+++ src/main/java/subway/application/LineService.java
@@ -5,15 +5,15 @@ import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
 import org.springframework.stereotype.Service;
-import subway.domain.Line;
-import subway.domain.Station;
+import subway.domain.line.Line;
+import subway.domain.line.Station;
 import subway.domain.Subway;
-import subway.dto.LineRequest;
-import subway.dto.LineResponse;
-import subway.dto.LineSelectResponse;
-import subway.dto.LinesSelectResponse;
-import subway.dto.StationSaveRequest;
-import subway.dto.StationSelectResponse;
+import subway.dto.line.LineRequest;
+import subway.dto.line.LineResponse;
+import subway.dto.line.LineSelectResponse;
+import subway.dto.line.LinesSelectResponse;
+import subway.dto.station.StationSaveRequest;
+import subway.dto.station.StationSelectResponse;
 import subway.exception.station.AlreadyExistStationException;
 import subway.exception.station.StationNotFoundException;
 import subway.repository.LineRepository;
