+++ src/main/java/subway/application/PathService.java
@@ -3,13 +3,13 @@ package subway.application;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
-import subway.domain.Fare;
-import subway.domain.FarePolicy;
-import subway.domain.Station;
+import subway.domain.fare.Fare;
+import subway.domain.fare.FarePolicy;
+import subway.domain.line.Station;
 import subway.domain.Subway;
 import subway.domain.SubwayGraph;
-import subway.dto.ShortestPathSelectResponse;
-import subway.dto.StationSelectResponse;
+import subway.dto.path.ShortestPathSelectResponse;
+import subway.dto.station.StationSelectResponse;
 import subway.exception.station.StationNotFoundException;
 import subway.repository.LineRepository;
 import subway.repository.StationRepository;
