+++ src/main/java/subway/domain/Subway.java
@@ -3,6 +3,9 @@ package subway.domain;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
+import subway.domain.line.Line;
+import subway.domain.line.Section;
+import subway.domain.line.Station;
 import subway.exception.DuplicatedNameException;
 import subway.exception.line.NonExistLineException;
