+++ src/main/java/subway/domain/fare/FarePolicy.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.fare;
 
 import org.springframework.stereotype.Component;
