+++ src/main/java/subway/domain/line/Section.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.line;
 
 import java.util.Objects;
 import subway.exception.section.InvalidSectionDirectionException;
