+++ src/main/java/subway/dto/line/LineResponse.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.dto.line;
 
 import subway.entity.LineEntity;
