+++ src/main/java/subway/dto/station/StationSaveRequest.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.dto.station;
 
 public class StationSaveRequest {
