+++ src/main/java/subway/dto/station/StationSelectResponse.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.dto.station;
 
 import subway.entity.StationEntity;
