+++ src/main/java/subway/entity/StationEntity.java
@@ -3,8 +3,8 @@ package subway.entity;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
-import subway.domain.Line;
-import subway.domain.Station;
+import subway.domain.line.Line;
+import subway.domain.line.Station;
 
 public class StationEntity {
