+++ src/main/java/subway/repository/LineRepository.java
@@ -9,9 +9,9 @@ import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Repository;
-import subway.domain.Line;
-import subway.domain.Section;
-import subway.domain.Station;
+import subway.domain.line.Line;
+import subway.domain.line.Section;
+import subway.domain.line.Station;
 import subway.domain.Subway;
 import subway.entity.LineEntity;
 import subway.entity.SectionEntity;
