+++ src/test/java/subway/domain/DistanceTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
+import subway.domain.line.Distance;
 import subway.exception.section.NegativeDistanceValueException;
 
 @DisplayNameGeneration(ReplaceUnderscores.class)
