+++ src/test/java/subway/domain/SectionTest.java
@@ -8,6 +8,8 @@ import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import subway.domain.line.Section;
+import subway.domain.line.Station;
 import subway.exception.section.InvalidSectionDirectionException;
 
 @DisplayNameGeneration(ReplaceUnderscores.class)
