+++ src/test/java/subway/domain/StationTest.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
+import subway.domain.line.Station;
 import subway.exception.EmptyNameException;
 
 @DisplayNameGeneration(ReplaceUnderscores.class)
