+++ src/test/java/subway/domain/SubwayTest.java
@@ -7,6 +7,9 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
+import subway.domain.line.Line;
+import subway.domain.line.Section;
+import subway.domain.line.Station;
 import subway.exception.DuplicatedNameException;
 import subway.exception.line.NonExistLineException;
 import subway.exception.station.StationNotFoundException;
