+++ src/test/java/subway/repository/StationRepositoryTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import subway.domain.Station;
+import subway.domain.line.Station;
 import subway.repository.dao.StationDao;
 
 @DisplayNameGeneration(ReplaceUnderscores.class)
